+++ QMath/QMath/Sequences.cpp
@@ -1,7 +1,7 @@
 #include "Sequences.h"
 
 namespace QMath {
-	int Sequences::getFibonacciNumber(int n) {
+	long Sequences::getFibonacciNumber(long n) {
 		return n == 1 ? 1 : n == 2 ? 2 : getFibonacciNumber(n - 2) + getFibonacciNumber(n - 1);
 	}
 }
