+++ QMath/QMath/Sums.h
@@ -1,10 +1,8 @@
 #pragma once
 
-#include <list>
-
 namespace QMath {
 	class Sums {
 	public:
-		static int getSum(std::list<int> list);
+		static long getSum(std::list<long> list);
 	};
 }
